fix: resolve webpack runtime error by isolating server-only utilities
