+++ lib/utils/server.ts
@@ -1,0 +1,111 @@
+import { getSelectedModel } from '@/lib/actions/users'
+import { createOpenAI } from '@ai-sdk/openai'
+import { createGoogleGenerativeAI } from '@ai-sdk/google'
+import { createAmazonBedrock } from '@ai-sdk/amazon-bedrock'
+import { createXai } from '@ai-sdk/xai'
+
+export async function getModel(requireVision: boolean = false) {
+  const selectedModel = await getSelectedModel();
+
+  const xaiApiKey = process.env.XAI_API_KEY;
+  const gemini3ProApiKey = process.env.GEMINI_3_PRO_API_KEY;
+  const awsAccessKeyId = process.env.AWS_ACCESS_KEY_ID;
+  const awsSecretAccessKey = process.env.AWS_SECRET_ACCESS_KEY;
+  const awsRegion = process.env.AWS_REGION;
+  const bedrockModelId = process.env.BEDROCK_MODEL_ID || 'anthropic.claude-3-5-sonnet-20241022-v2:0';
+  const openaiApiKey = process.env.OPENAI_API_KEY;
+
+  if (selectedModel) {
+    switch (selectedModel) {
+      case 'Grok 4.2':
+        if (xaiApiKey) {
+          const xai = createXai({
+            apiKey: xaiApiKey,
+            baseURL: 'https://api.x.ai/v1',
+          });
+          try {
+            const modelId = requireVision ? 'grok-vision-beta' : 'grok-2-1212';
+            return xai(modelId);
+          } catch (error) {
+            console.error('Selected model "Grok 4.2" is configured but failed to initialize.', error);
+            throw new Error('Failed to initialize selected model.');
+          }
+        } else {
+            console.error('User selected "Grok 4.2" but XAI_API_KEY is not set.');
+            throw new Error('Selected model is not configured.');
+        }
+      case 'Gemini 3':
+      case 'Gemini 3.1 Pro':
+        if (gemini3ProApiKey) {
+          const google = createGoogleGenerativeAI({
+            apiKey: gemini3ProApiKey,
+          });
+          try {
+            return google('gemini-1.5-pro-latest');
+          } catch (error) {
+            console.error('Selected model "Gemini 3.1 Pro" is configured but failed to initialize.', error);
+            throw new Error('Failed to initialize selected model.');
+          }
+        } else {
+            console.error('User selected "Gemini 3.1 Pro" but GEMINI_3_PRO_API_KEY is not set.');
+            throw new Error('Selected model is not configured.');
+        }
+      case 'GPT-5.1':
+        if (openaiApiKey) {
+          const openai = createOpenAI({
+            apiKey: openaiApiKey,
+          });
+          return openai('gpt-4o');
+        } else {
+            console.error('User selected "GPT-5.1" but OPENAI_API_KEY is not set.');
+            throw new Error('Selected model is not configured.');
+        }
+    }
+  }
+
+  // Default behavior: Grok -> Gemini -> Bedrock -> OpenAI
+  if (xaiApiKey) {
+    const xai = createXai({
+      apiKey: xaiApiKey,
+      baseURL: 'https://api.x.ai/v1',
+    });
+    try {
+      const modelId = requireVision ? 'grok-vision-beta' : 'grok-latest';
+      return xai(modelId);
+    } catch (error) {
+      console.warn('xAI API unavailable, falling back to next provider:');
+    }
+  }
+
+  if (gemini3ProApiKey) {
+    const google = createGoogleGenerativeAI({
+      apiKey: gemini3ProApiKey,
+    });
+    try {
+      return google('gemini-1.5-pro-latest');
+    } catch (error) {
+      console.warn('Gemini 3.1 Pro API unavailable, falling back to next provider:', error);
+    }
+  }
+
+  if (awsAccessKeyId && awsSecretAccessKey) {
+    const bedrock = createAmazonBedrock({
+      bedrockOptions: {
+        region: awsRegion,
+        credentials: {
+          accessKeyId: awsAccessKeyId,
+          secretAccessKey: awsSecretAccessKey,
+        },
+      },
+    });
+    const model = bedrock(bedrockModelId, {
+      additionalModelRequestFields: { top_k: 350 },
+    });
+    return model;
+  }
+
+  const openai = createOpenAI({
+    apiKey: openaiApiKey,
+  });
+  return openai('gpt-4o');
+}
